+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog for Flamenco Manager
 ==============================
 
+## Version 2.0.15 (in development)
+
+- Flamenco Manager can now be run from a different directory than the executable is in. It searches
+  for web templates both relative to the current working directory and relative to the executable's
+  own directory.
+
+
 ## Version 2.0.14 (released 2017-09-07)
 
 - Fixed panic when enabling UPnP/SSDP auto-discovery on Windows.